refactor(viewer): organize application modules
